+++ CHANGELOG.md
@@ -36,10 +36,12 @@
 * `axios` installed with version `0.24.0`.
 * `bootstrap` installed with version `5.1.3`.
 * `fuse.js` installed with version `6.4.6`.
+* `image-search-google` installed with version `1.3.0`.
 * `qs` installed with version `6.10.1`.
 * `sass-loader` installed with version `12.3.0`.
 * `sass` installed with version `1.43.4`.
 * `sweetalert2` installed with version `11.2.0`.
+* `uniqid` installed with version `5.4.0`.
 * `v-tooltip` installed with version `4.0.0-beta.2`.
 * `vee-validate` installed with version `4.5.6`.
 * `vue-good-table-next` installed with version `0.1.0`.
@@ -50,3 +52,5 @@
 * `vue-select` installed with version `4.0.0-beta.1`.
 * `vue-toastification` installed with version `2.0.0-rc.5`.
 * `xss` installed with version `1.0.10`.
+* `yup` installed with version `0.32.11`.
+* `yup-locales` installed with version `0.32.11`.